ClientManager implemented fully. Client implemented halfway until configureReceives-method
Changed the return type of registerReceiver-method of ClientManager to Boolean. Since it receives a Boolean from registerReceiver-method of DatabaseWrapper and has to inform Client. Cleaner and faster alternative instead of implementing a new method in Client.

Review needed for callbacks in configureReceives-method in Client. Unsure if correct. onReceive for Event.REFRESH_SLOT_APPROX needs review because unsure if receivers.get(slotCode) should be non-nullable(!!).